Backend_todo/requirements.txt
Flask==3.0.2
requests
psycopg2-binary
nats-py 
aiohttp